+++ js/main.js
@@ -1,18 +1,13 @@
-import { photos } from './setup.js';
-import { container, generateThumbnails } from './thumbs.js';
-import { openBigPicture } from './full-size-pictures.js';
+// import { photos } from './setup.js';
+
+// import { openBigPicture } from './full-size-pictures.js';
 import './form.js';
 import './img-effects.js';
 import './scale-control.js';
 import './server-data.js';
+import { renderRemoteThumbnails, openCard } from './server-data.js';
+
+renderRemoteThumbnails();
+openCard();
 
 
-generateThumbnails(photos);
-
-container.addEventListener('click', (evt) => {
-  const currentPhoto = evt.target.closest('.picture');
-
-  if (currentPhoto) {
-    openBigPicture(currentPhoto.dataset.pictureId);
-  }
-});
